feat(DTFS2-7393): code review on e2e tests
